Use session cookie for anonymous CSRF cookie.
This fixes an issue with IE 7 and 8 not accepting the cookie if the client's
date is set wrong. Session cookies (cookies that have no set expiry date and
therefore expire at the end of the browser session) are already used by Django
for storing the SESSION_COOKIE (sessionid by default).