Remove token once used. Generate new token when template tag {{csrf_token}} is actually used
Changed the behaviour too 'one time' only tokens.

Once a form is POSTed remove the token in the session.
Generating a new token is done when the template tag {{csrf_token}} is
actully used in a template.

With the setting 'CSRF_REMOVE_USED_TOKENS' you can override the 'remove
token after post' behaviour.